allow using different Authentication Database then the analysed one
In some cases (particularly when connecting to an Atlas cluster) the
Authentication Database will be different then the one we want to
analyse. In those cases the mongo command of form:

```
mongo target_db_name —-host “some_host/?authSource=auth_db_name”
```

which is generated by variety-cli will result in a connection error

some claim it should be possible to specify [different auth database and
target database](https://dba.stackexchange.com/questions/185783/how-to-specify-authentication-database-and-target-database-separately-on-mongodb) however none of the connection string
formats I am aware of worked for me with Atlas

This commit uses `db.getSiblingDB` (added to the `eval` string) to
switch to the desired database once connection is established. The
`target_db_name ` argument of the mongo command is removed.